🪲 [Fix]: Fix linter settings and docs (#4)
This pull request includes minor configuration and documentation
improvements related to linting and module test documentation. The main
changes are the addition of a description to the `Get-PSModuleTest.ps1`
function, updates to markdown linting rules, and the disabling of
several linter validations in the GitHub Actions workflow.

Linter and workflow configuration updates:

* Disabled several linter validations in the `Linter` section of
`.github/PSModule.yml` by setting related environment variables to
`false`.
* Updated `.github/linters/.markdown-lint.yml` to turn off the `MD024`
(no-duplicate-heading) rule and made a minor formatting fix for the
`MD026` punctuation list.

Documentation improvements:

* Added a `.DESCRIPTION` section to the `Get-PSModuleTest.ps1` function
to provide more context in the help output.